\name{bayessir-package}
\alias{bayessir-package}
\docType{package}
\title{
PMMH algorithm for time-varying SIRS model
}
\description{
Uses a particle marginal Metropolis-Hastings (PMMH) algorithm to sample from the posterior distribution of the environmental and transmission parameters of a hidden SIRS model given observed cholera data.
}
\details{
\tabular{ll}{
Package: \tab bayessir\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-01-09\cr
License: \tab GPL-3 \cr
}

Main algorithm and example: \code{\link{bayessir}}
}

\author{
Amanda Koepke and Vladimir Minin

Maintainer: Amanda Koepke <alallen@stat.washington.edu>
}
\references{
Andrieu, Christophe, Arnaud Doucet, and Roman Holenstein. "Particle Markov chain Monte Carlo methods." Journal of the Royal Statistical Society: Series B (Statistical Methodology) 72.3 (2010): 269-342.
}


\keyword{ package }
