% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inhomoModPoissonTLforSIWR}
\alias{inhomoModPoissonTLforSIWR}
\title{Modified inhomogeneous Poisson tau-leaping algorithm for SIWR model}
\usage{
inhomoModPoissonTLforSIWR(startState, popSize, Wmax, startTime, intervalLength,
  parameters, alphas, allbreaks, deltatint, ncrit)
}
\arguments{
\item{startState}{Integer vector of size three containing the initial values for S, I, and W}

\item{popSize}{Population size}

\item{startTime}{Start time of simulation}

\item{intervalLength}{Length of simulation interval}

\item{parameters}{Numeric vector of size six containing the current values of the infectious contact rate, betaW, the recovery rate, the rate at which immunity is lost, kappa, and eta}

\item{alphas}{Numeric vector containing the values of alpha}

\item{allbreaks}{Numeric vector containing the times at which the value of alpha changes}

\item{deltatint}{Initial value to use for tau in tau-leaping algorithm}

\item{ncrit}{Critical size for modified tau-leaping algorithm; if the population of a compartment is lower than this number a single step algorithm is used until the population gets above the critical size.}
}
\value{
Vector containing values of S, I, and W at the end of the simulation interval
}
\description{
Uses a tau-leaping algorithm to simulate S, I, W, and R forward in time when value of environmental force of infection changes
}

