% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ModPoissonTLforSIWR}
\alias{ModPoissonTLforSIWR}
\title{Modified homogeneous Poisson tau-leaping algorithm for SIWR model}
\usage{
ModPoissonTLforSIWR(startState, popSize, Wmax, startTime, intervalLength,
  parameters, alpha, deltatint, ncrit)
}
\arguments{
\item{startState}{Integer vector of size three containing the initial values for S, I, and W}

\item{popSize}{Population size}

\item{startTime}{Start time of simulation}

\item{intervalLength}{Length of simulation interval}

\item{parameters}{Numeric vector of size six containing the current values of the infectious contact rate, betaW, the recovery rate, the rate at which immunity is lost, kappa, and eta}

\item{alpha}{The value of the time-varying environmental force of infection}

\item{deltatint}{Initial value to use for tau in tau-leaping algorithm}

\item{ncrit}{Critical number}
}
\value{
Vector containing values of S, I, and W at the end of the simulation interval
}
\description{
Modified Poisson tau-leaping algorithm. Uses a tau-leaping algorithm to simulate S, I, W, and R forward in time when value of environmental force of infection remains constant
}

