% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampleOnce}
\alias{sampleOnce}
\title{Samples a vector based on weights}
\usage{
sampleOnce(weights)
}
\arguments{
\item{weights}{input numeric vector}
}
\value{
Index of vector selected
}
\description{
Samples a vector based on weights
}

