% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SMCforSIWR}
\alias{SMCforSIWR}
\title{Sequential Monte Carlo algorithm for SIWR}
\usage{
SMCforSIWR(observedCounts, observedTimes, parameters, alphaBreaks, startMeans,
  nparticles, popSize, reportingRate, deltatval, ncrit, UseGill, maxWval)
}
\arguments{
\item{observedCounts}{Integer vector of observed counts}

\item{observedTimes}{Numeric vector of observation times}

\item{parameters}{Numeric vector of size seven containing the current values of the infectious contact rate, the recovery rate, the rate at which immunity is lost, and the powers}

\item{alphaBreaks}{List containing the time-varying environmental force of infection and the times at which the rate changes, broken up by simulation intervals}

\item{startMeans}{Numeric vector of size two containing the means of initial distributions for the numbers of susceptible and infected individuals respectively}

\item{nparticles}{Number of particles}

\item{popSize}{Population size}

\item{reportingRate}{Probability of infected individuals seeking treatment}

\item{deltatval}{Initial value to use for tau in tau-leaping algorithm}

\item{ncrit}{Critical size for modified tau-leaping algorithm; if the population of a compartment is lower than this number a single step algorithm is used until the population gets above the critical size.}

\item{UseGill}{boolian; if 1, uses the gillespie algorithm. If 0, uses tau-leaping algorithm}

\item{maxWval}{Upper bound for W compartment. If 0, no bounding.}
}
\value{
log-likelihood and estimate of last state
}
\description{
Sequential Monte Carlo algorithm for SIWR
}

